5218: Restore exclusions of some dependencies
I removed all the exclusions in the maven config because it was annoying
having to copy extra jar files into Tomcat to start the application. It
turns out that some of the exclusions were actually necessary. For
example, having both log4j-over-slf4j.jar and slf4j-log4j12.jar on the
classpath [causes a StackOverflow](http://www.slf4j.org/codes.html#log4jDelegationLoop).

To resolve these issues I'm partially reverting the previous commit,
which adds in all the exclusions that used to be there, except the ones
which seem directly related to the .jar files that you used to have to
copy from AdditionalFiles.

Authors: @richardtowers