5218: Ensure all logs go to STDOUT
When deploying to PaaS it's difficult to read files that have been
written to disk, and they don't persist if an instance needs to restart.
It's much more convenient to log everything to STDOUT.

This commit removes all the rolling file logger specific stuff and
switches to using the ConsoleAppender for everything.

Authors: @richardtowers @niyioyelade