5218: Split manifest files for different applications
Because we're deploying the EidasNode to two different servers, once as
a ConnectorNode and once as a ProxyNode, we need to configure it
differently. This means we need different environment variables, which
at the moment means different manifest files. There's probably a nicer
way of achieving this, but we can cross that bridge later.

Authors: @richardtowers @niyioyelade