5220: Switch to using Metadata signing keys everywhere
The reference implementation appears to use the same key for signing
Metadata and SAML messages. Originally we were allowing it to use the
key intended for signing SAML for Metadata too, but it's better if it's
the other way around. This commit switches to use Metadata siging keys
for signing Metadata and SAML. This should allow us to use the Metadata
CA as the main trust anchor shared with third parties.

Also in this commit we've switched from (confusingly) using a signing
key in the EncryptModule_IdP.xml config file. The IdP shouldn't need an
encryption key, because nothing it interacts with should be sending it
encrypted assertions or messages. The implementation expects one though.
Previously we hadn't generated a key for this and were just bodging it
with a signing key so that it wouldn't complain. Better to generate an
encryption key and configure it to use that (even if it isn't going to
be doing any decryption).

Authors: @richardtowers @niyioyelade