5220: Use environment variables for configuring Node Keystores
The previous approach of hard-configuring distinguished names and serial
numbers was bad for a couple of reasons:

1) We had to duplicate these settings in several files (now they just
live where the environment variables are set)
2) We had to have a horrible find and replace that needed to know about
the values.
3) It was painful to deploy this to our Platform as a Service
environment because we would have had to propagate the find-and-replace
hack into more production-like scripts.

We still have to do some find and replacing because when we run locally
both instances of the Node run in the same Tomcat instance. This means
they get the same environment variables, so we can't use these to
distinguish between the nodes. A bit hacky, but a bit less hacky than
before so I don't feel as bad about it as I should...

Authors: @richardtowers