5220: Switch back to using the test keystore when running locally
No need to use fancy keystores when running locally, it's good enough to
use the same key / cert in all the applications.

Now that we have environment variables for all the config we can
configure the PaaS deployed versions to use the correct keystores just
using the environment vars.

Authors: @richardtowers