5223: Fix configuration to support authn requests from the bridge
There were a couple of misconfigurations when running the reference code
locally. The metadata served by the connector node was incorrectly
telling the bridge to send AuthnRequests to the proxy node. To fix this
we've had to add a copy of the EIDAS-Config folder that we can point the
proxy node at. This allows us to put correct config for the connector
node in EIDAS-Config and correct config for the proxy node in
EIDAS-Config-Proxy-Node. Feels pretty hacky though because we've
duplicated loads of config.

Also added the certificate of the bridge to the connector node's
keystore (note - the key the bridge is using is from the ida-dev-pki
which maybe is not the right thing to do).

Also gitignored some eclipse files for Suneet.

Also added jpda to catalina run, since this script is only used to run
locally and we usually want the ability to attach a debugger.

Authors: @richardtowers @sfkamath