5221: Deploy to multiple tomcats when running locally
We had no end of trouble getting all four of the applications to run on
the same instance of tomcat with our environment-variable based config.
Because there's only one instance all the applications end up with the
same environment variables, which meant that we had to have different
config files for each application.

On our "production" system each application runs in its own process, so
we can distinguish them just using environment variables. By using
separate tomcats locally we can use the same approach.

This has allowed us to remove the (hacky) EIDAS-Config-Proxy-Node
folder, and the old tomcat-deploy.sh script (which was also hacky).

See the changes to the README for an overview of how to run the
applications now.

Also: fixed log4j.xml so that stub-sp writes its logs to STDOUT.

Authors: @richardtowers @sfkamath